requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
amqp==5.1.1
anyio==3.5.0
argcomplete==2.0.0
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
bech32==1.2.0
billiard==3.6.4.0
black==22.3.0
boto3==1.24.17
botocore==1.27.17
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cosmospy==6.0.0
cryptography==38.0.1
databases==0.5.5
datamodel-code-generator==0.13.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.75.2
flower==1.1.0
frozenlist==1.3.0
genson==1.2.2
gunicorn==20.1.0
h11==0.13.0
hdwallets==0.1.2
HTMLParser==0.0.2
httptools==0.4.0
humanize==4.2.3
hvac==1.0.0
idna==3.3
inflect==5.6.0
isodate==0.6.1
isort==5.10.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
kombu==5.2.4
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
mnemonic==0.20
motor==3.0.0
multidict==6.0.2
mypy-extensions==0.4.3
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
prance==0.21.8.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pycparser==2.21
pydantic==1.9.0
pyhcl==0.4.4
PyJWT==2.3.0
pymongo==4.1.1
pyrsistent==0.16.1
PySnooper==1.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
PyYAML==6.0
qbittorrent-api==2022.2.28
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
typed-ast==1.5.4
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
yarl==1.7.2
